LayoutTests/sputnik/Conformance/12_Statement/12.2_Variable_Statement/S12.2_A8_T1-expected.txt

CONSOLE MESSAGE: line 78: SyntaxError: Unexpected token '+='. Expected ';' after variable declaration.
S12.2_A8_T1

PASS Expected parsing failure

TEST COMPLETE
